fix(common): make mixins work without TSC errors
affects: @blaze-elements/common

ISSUES CLOSED: #303, #300
