+++ packages/common/components.ts
@@ -1,0 +1,4 @@
+import { Component } from 'skatejs';
+import { Colored, Disabled } from './mixins';
+
+export const ColoredDisabledComponent = Colored( Disabled( Component ) );
